Fixing circular dependency issue
